Initial version of spichipd
This preliminary test version compiles with the Makefile in src/ directory.

For 'daemon' operation systemd is needed, search from web 'how to write
modern systemd daemon'.

Documentation is in comment blocks of the C++ header and code files.
The 'Doxyfile' allows to generate the class documents with 'doxygen'
command though this will probably have to be tuned more later.